backend/requirements.txt
fastapi[all]
sqlalchemy
uvicorn[standard]
python-dotenv
PyPdf2
langchain
langchain_openai